requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aspose-cells==22.11.0
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.57.0
gspread==5.7.2
httplib2==0.21.0
idna==3.4
JPype1==1.4.1
multidict==6.0.2
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
packaging==22.0
panda==0.3.1
pandas==1.5.2
protobuf==4.21.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.13
XlsxWriter==3.0.3
yarl==1.8.1
